requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.10
sqlparse==0.5.1
django-environ==0.11.2
## The following requirements were added by pip freeze:
djangorestframework-simplejwt==5.3.1
pillow==11.1.0
PyJWT==2.10.1
rest-framework-simplejwt==0.0.2